Move environment utilities, more descriptive ts-is property test name
