+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./dateDeadline";
 export * from "./dialog";
+export * from "./env";
 export * from "./utils";
